+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `soyhuce/next-ide-helper` will be documented in this file
 
 ## [Next release] - YYYY-MM-DD
 
+### Added
+
+- Add ability to override type for model relations
+
 ## [0.4.1] - 2021-02-16
 
 ### Added